+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Title+Package Record: Display Usage & Analysis Accordion. (UIEH-949)
 * Added Usage Consolidation Summary table to Package Show page. (UIEH-945)
 * Added Usage Consolidation Filters to Resource Show page. (UIEH-951)
+* Title Record: Display Usage & Analysis Accordion. (UIEH-950)
 
 ## [5.0.0] (https://github.com/folio-org/ui-eholdings/tree/v5.0.0) (2020-10-15)